refactor(config): extract ProviderKind into a provider_kind module (#3311)
Second behavior-preserving slice of the config module split. Moves the
ProviderKind enum (with its serde aliases) and its identity impl
(ALL / all / names_hint / as_str / parse / is_siliconflow / provider)
verbatim out of lib.rs into a focused `provider_kind` module, re-exported
`pub use provider_kind::ProviderKind` at the crate root so
`codewhale_config::ProviderKind` and all `crate::ProviderKind` paths are
unchanged. The module depends only on `crate::provider`; no values or
logic changed.

lib.rs shrinks by ~158 lines (~4.38k now), continuing to separate provider
identity from config schema/loading per #3311 / #2608.

Verified: cargo fmt --check, cargo clippy --all-targets,
cargo test -p codewhale-config (248 passed, unchanged).
